탈퇴 수정

+++ HoneyJamGuestHouse/src/main/java/com/honeyjam/member/servlet/DeleteMemberServlet.java
@@ -6,6 +6,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import com.honeyjam.exception.LoginFailException;
 import com.honeyjam.member.service.MemberService;
@@ -24,9 +25,17 @@ public class DeleteMemberServlet extends HttpServlet {
 			String password = request.getParameter("password");
 			
 			try {
+				HttpSession session = request.getSession();
 				MemberService service= MemberServiceImpl.getInstance();
+				
+				if(!email.equals(((Member)session.getAttribute("loginMember")).getEmail())) {
+					throw new LoginFailException("현재 로그인한 이메일로 다시 입력하세요");
+				}
+				
 				Member member = service.authenticate(email, password);
 				
+				
+				session.invalidate();
 				service.removeMemberById(email);
 				response.sendRedirect(getServletContext().getInitParameter("rootPath")+"/main.jsp");
